Adjusted price value... should be double

+++ java/controller/AdministratorCreateTransitController.java
@@ -21,7 +21,7 @@ public class AdministratorCreateTransitController {
     public void create() throws SQLException {
         if (!route.getText().isEmpty() ||  !price.getText().isEmpty() || !type.getSelectionModel().isEmpty()) {
             try {
-                TransitDAO.createTransit(route.getText(), type.getSelectionModel().getSelectedItem().toString(), Integer.parseInt(price.getText()));
+                TransitDAO.createTransit(route.getText(), type.getSelectionModel().getSelectedItem().toString(), Double.parseDouble(price.getText()));
             }
             catch(SQLException e) {
                 System.out.println("Issue with SQL");
